Instead of having an external script, make it an included one
Then the CLI will just invoke it after downloading the command.
Simplifies things a bit too.

Then also make it an `edb` tool, so you can run `edb load-ext` in
devmode.  That was actually the most annoying part of this.  (But only
because I wanted the edb tool to use the options directly, I guess.)
